backend/requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
keras==2.11.0
matplotlib==3.7.1
mmcv_full==1.4.0
numpy==1.23.3
opencv_contrib_python==4.5.5.62
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==1.5.1
Pillow==9.2.0
python-dotenv==1.0.0
PyYAML==6.0
Requests==2.31.0
scikit_image==0.19.3
scipy==1.10.1
tensorflow==2.11.0
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm==4.64.1
#mmdet.egg==info
